# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health-check/{customer_id}":{"get":{"summary":"Health Check","operationId":"health_check_health_check__customer_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}},{"name":"database","in":"query","required":true,"schema":{"type":"string","title":"Database"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/items/{customer_id}":{"get":{"summary":"Read Items","operationId":"read_items_items__customer_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}},{"name":"database","in":"query","required":true,"schema":{"type":"string","title":"Database"}},{"name":"table","in":"query","required":true,"schema":{"type":"string","title":"Table"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Put Item","description":"Put an item in **database** using table **table** with values (**id**, **value**)\n\nNOTE: Works only for tables with a primary key (id) and an additional key (with value **value**)","operationId":"put_item_items__customer_id__put","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}},{"name":"database","in":"query","required":true,"schema":{"type":"string","title":"Database"}},{"name":"table","in":"query","required":true,"schema":{"type":"string","title":"Table"}},{"name":"id","in":"query","required":true,"schema":{"type":"integer","title":"Id"}},{"name":"value","in":"query","required":true,"schema":{"type":"string","title":"Value"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Item","description":"Delete an item from **database** using table **table** with primary key **id_key** and id **id**\n\nNOTE: Works only for tables with a primary key (id)","operationId":"delete_item_items__customer_id__delete","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}},{"name":"database","in":"query","required":true,"schema":{"type":"string","title":"Database"}},{"name":"table","in":"query","required":true,"schema":{"type":"string","title":"Table"}},{"name":"id_key","in":"query","required":true,"schema":{"type":"string","title":"Id Key"}},{"name":"id","in":"query","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/item/{customer_id}/{id}":{"get":{"summary":"Read Item","operationId":"read_item_item__customer_id___id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}},{"name":"database","in":"query","required":true,"schema":{"type":"string","title":"Database"}},{"name":"table","in":"query","required":true,"schema":{"type":"string","title":"Table"}},{"name":"key","in":"query","required":true,"schema":{"type":"string","title":"Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cluster":{"get":{"summary":"Read Cluster Status","operationId":"read_cluster_status_cluster_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/cluster/{customer_id}":{"get":{"summary":"Read Customer Cluster Status","operationId":"read_customer_cluster_status_cluster__customer_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Register Cluster Node","operationId":"register_cluster_node_cluster__customer_id__put","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}},{"name":"node_ip","in":"query","required":true,"schema":{"type":"string","title":"Node Ip"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cluster/{node_ip}":{"delete":{"summary":"Register Cluster Node","description":"Normally, there'd be an intermediary keeping a mapping of customer_id to\nnode_ip which would talk to this API acting as a sort of loadbalancer (healthcheck wise)\n\nIn absence of such middleware, we simply assume IP is unique per customer_id (cluster).\nThus, removing an IP would not remove any other node.","operationId":"register_cluster_node_cluster__node_ip__delete","parameters":[{"name":"node_ip","in":"path","required":true,"schema":{"type":"string","title":"Node Ip"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}